e2e: Add TaskRun for running kind tests.
This adds a TaskRun for executing kind tests. To start, this is a single
TaskRun, but I plan to break this out into its own reusable Task for the
catalog in the future.

The configs here are heavily based on the configs in
https://github.com/kind-ci/examples and
https://github.com/kubernetes-sigs/kind/issues/303.

This was done in favor of the Prow based config, since the cluster
failed to come up, dispite the various labels set (likely the settings
weren't applying / we are missing some other config in our prow
cluster). Since we want to move towards Tekton based CI in the future
anyway, it was simpler to handle container setup in a Task from scratch.

Much of the component installation is unfortunately copied verbatim
from the install script - there is some strange interaction going on
between kind and ko that causes ko to fail if called or sourced
from e2e.sh. I plan to investigate this further in another PR.