Move dynamic watcher params to ConfigMap.
This change moves dynamic params for watcher - e.g. any type of parameter that we
expect users will want to turn on/off or modify freely during the
lifetime of the process - to a ConfigMap.

For now, this only includes the update annotation flag, but we want to
use this for grace period flags included for Run cleanup.
This is particularly useful to modify configs freely for testing
w/o modifying the upstream deploy yamls.